refactor: move api related services to api subfolder
