+++ src/cli/template-deploy.ts
@@ -7,16 +7,16 @@ import { DeploymentTypePrompt } from '../core/prompt/deployment-type.prompt.js';
 import { OrganizationPrompt } from '../core/prompt/organization.prompt.js';
 import { TemplatePrompt } from '../core/prompt/template.prompt.js';
 import { WorkspacePrompt } from '../core/prompt/workspace.prompt.js';
-import { ApiService } from '../core/service/api.service.js';
-import { AuthService } from '../core/service/auth.service.js';
+import { ApiService } from '../core/service/api/api.service.js';
+import { AuthService } from '../core/service/api/auth.service.js';
+import { FileService } from '../core/service/api/file.service.js';
+import { OrganizationService } from '../core/service/api/organization.service.js';
+import { TemplateService } from '../core/service/api/template.service.js';
+import { WorkspaceService } from '../core/service/api/workspace.service.js';
 import { ConfigService } from '../core/service/config.service.js';
 import { DeploymentService } from '../core/service/deployment.service.js';
-import { FileService } from '../core/service/file.service.js';
-import { OrganizationService } from '../core/service/organization.service.js';
 import { TaskService } from '../core/service/task.service.js';
-import { TemplateService } from '../core/service/template.service.js';
 import { ValidationService } from '../core/service/validation.service.js';
-import { WorkspaceService } from '../core/service/workspace.service.js';
 import { CreateTemplateCtx } from '../shared/ctx/create-template.ctx.js';
 import { DeploymentScope } from '../shared/enum/deployment-scope.enum.js';
 import { DeploymentType } from '../shared/enum/deployment-type.enum.js';
